% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.r
\name{myttest}
\alias{myttest}
\title{Perform a t-test or Welch's test on two samples}
\usage{
myttest(x, y, paired = FALSE, alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector representing the first sample}

\item{y}{A numeric vector representing the second sample}

\item{paired}{A logical value indicating whether the samples are paired (default: FALSE)}

\item{alpha}{The significance level for hypothesis testing (default: 0.05)}
}
\value{
A list with the following components:
\itemize{
  \item test_type: A character string indicating the test performed ("T-test", "Welch", or "Paired")
  \item reject_null: A logical value indicating whether the NULL hypothesis should be rejected
  \item summary: A summary object containing the results of the appropriate t.test function
  \item data: A list containing the input samples (x and y)
}
}
\description{
This function performs a t-test or Welch's test on two samples to test the NULL hypothesis
that the underlying means of the two populations are the same. It takes into consideration
the variances and whether the samples are paired or not.
}
\examples{
set.seed(32)
x <- rnorm(30, mean = 10, sd = 15)
y <- rnorm(30, mean = 8, sd = 15)
ans1 <- myttest(x, y, alpha = 0.05, paired = FALSE)
print(ans1)
plot(ans1)

}
