% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.r
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plot method for Rttest objects}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{An object of class "Rttest" returned by the myttest function}

\item{...}{Additional arguments passed to the plot function (currently unused)}
}
\description{
This function creates boxplots for non-paired data or boxplot of differences for paired data,
and includes the confidence interval for the difference of means when plotting.
}
\examples{
set.seed(32)
x <- rnorm(30, mean = 10, sd = 15)
y <- rnorm(30, mean = 8, sd = 15)
ans1 <- myttest(x, y, alpha = 0.05, paired = FALSE)
plot(ans1)
}
