% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.r
\name{print.Rttest}
\alias{print.Rttest}
\title{Print method for Rttest objects}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{An object of class "Rttest" returned by the myttest function}

\item{...}{Additional arguments passed to the print function (currently unused)}
}
\description{
This function prints the results of an Rttest object in a human-readable format,
displaying the test type, whether the NULL hypothesis should be rejected, and
the confidence interval for the difference of means.
}
\examples{
set.seed(32)
x <- rnorm(30, mean = 10, sd = 15)
y <- rnorm(30, mean = 8, sd = 15)
ans1 <- myttest(x, y, alpha = 0.05, paired = FALSE)
print(ans1)
}
